+++ app/src/main/java/woowacourse/omok/data/repository/OmokRepositoryImpl.kt
@@ -1,22 +1,22 @@
 package woowacourse.omok.data.repository
 
-import woowacourse.omok.data.datasource.OmokDataSource
-import woowacourse.omok.data.db.toEntity
+import woowacourse.omok.data.dao.OmokDao
+import woowacourse.omok.data.db.omok.toEntity
 import woowacourse.omok.domain.point.Point
 import woowacourse.omok.domain.repository.OmokRepository
 
 class OmokRepositoryImpl(
-    private val omokDataSource: OmokDataSource,
+    private val omokDao: OmokDao,
 ) : OmokRepository {
     override fun saveNewPoint(newPoint: Point) {
-        omokDataSource.save(newPoint.toEntity())
+        omokDao.save(newPoint.toEntity())
     }
 
     override fun readAllPoint(): List<Point> {
-        return omokDataSource.readAll().map { it.toDomainModel() }
+        return omokDao.readAll().map { it.toDomainModel() }
     }
 
     override fun drop() {
-        omokDataSource.drop()
+        omokDao.drop()
     }
 }
